[Janik, Lea] Implement algorithm for distributing plants
The algorithm just works for one year at a time currently, although support for multiple generations can be added in the future.

Correct operation cannot be guaranteed, since seeding the database was and still is a proper Baustöh, therefore testing was difficult. Should probably work tho, since a lot of blood, sweat and tears went into designing the algorithm in the first place.